requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
sqlite3-python==1.0.0
python-dotenv==1.0.0
schedule==1.2.0
APScheduler==3.10.4
smtplib==3.11
Pillow==10.1.0
cloudscraper==1.2.71
retrying==1.3.4Flask==2.3.3
Flask-CORS==4.0.0